requirements.txt
snowflake-sqlalchemy
git+ssh://git@github.com/Daltix/product-team-tooling.git@develop#egg=credsman&subdirectory=credsman
